refactor: migrate Pydantic schemas into a structured package and add database documentation
