+++ schemas.py
@@ -1,23 +1,25 @@
 """
-schemas.py — Pydantic request/response models.
+schemas.py — DEPRECATED compatibility shim.
+
+All Pydantic schemas have moved to the schemas/ package:
+  schemas.auth     → RegisterRequest, LoginRequest, RefreshRequest, TokenResponse, UserResponse
+  schemas.evaluate → EvaluateRequest, EvaluateMultiRequest
+  schemas.document → UpdateAnalysisResultRequest
+  schemas.admin    → UpdateRoleRequest, UserSummary
+
+This file re-exports everything so any old imports keep working.
+Remove once all imports have been updated to use schemas.<module>.
 """
 
-from typing import List, Optional
-from pydantic import BaseModel
-
-
-class EvaluateRequest(BaseModel):
-    s3_object_key: Optional[str] = None
-    filename: Optional[str] = None
-
-
-class EvaluateMultiRequest(BaseModel):
-    s3_object_keys: List[str]
-    filenames: Optional[List[str]] = None
-    document_types: Optional[List[str]] = None
-    user_id: Optional[str] = None
-
-
-class UpdateAnalysisResultRequest(BaseModel):
-    score: Optional[int] = None
-    reasoning: Optional[str] = None
+from schemas import (  # noqa: F401
+    EvaluateMultiRequest,
+    EvaluateRequest,
+    LoginRequest,
+    RefreshRequest,
+    RegisterRequest,
+    TokenResponse,
+    UpdateAnalysisResultRequest,
+    UpdateRoleRequest,
+    UserResponse,
+    UserSummary,
+)
